reportes/requirements.txt
Flask
pandas
openpyxl
requests
weasyprint
python-dotenv